+++ src/main/java/com/nhnacademy/frontend/common/adapter/dto/book/response/BookCategoryNodeResponseDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.frontend.book.domain.response;
+package com.nhnacademy.frontend.common.adapter.dto.book.response;
 
 import java.util.List;
