+++ src/main/java/com/nhnacademy/frontend/common/service/BookService.java
@@ -1,7 +1,5 @@
 package com.nhnacademy.frontend.common.service;
 
-import com.nhnacademy.frontend.book.domain.response.BookCategoryNodeResponseDto;
-import com.nhnacademy.frontend.common.adapter.dto.book.response.SimpleBookResponseDto;
 import com.nhnacademy.frontend.common.adapter.dto.book.request.*;
 import com.nhnacademy.frontend.common.adapter.dto.book.response.*;
 import org.springframework.data.domain.Page;
