+++ src/main/java/com/nhnacademy/frontend/common/service/impl/BookServiceImpl.java
@@ -1,8 +1,6 @@
 package com.nhnacademy.frontend.common.service.impl;
 
 import com.nhnacademy.frontend.common.adapter.BookAdapter;
-import com.nhnacademy.frontend.book.domain.response.BookCategoryNodeResponseDto;
-import com.nhnacademy.frontend.common.adapter.dto.book.response.SimpleBookResponseDto;
 import com.nhnacademy.frontend.common.adapter.dto.book.request.*;
 import com.nhnacademy.frontend.common.adapter.dto.book.response.*;
 import com.nhnacademy.frontend.common.service.BookService;
