+++ src/test/java/com/nhnacademy/frontend/book/controller/AdminBookControllerTest.java
@@ -3,6 +3,7 @@ package com.nhnacademy.frontend.book.controller;
 import com.nhnacademy.frontend.admin.controller.AdminBookController;
 import com.nhnacademy.frontend.common.adapter.dto.book.response.BookResponseDto;
 import com.nhnacademy.frontend.auth.filter.JwtAuthenticationFilter;
+import com.nhnacademy.frontend.common.adapter.dto.book.response.SimpleBookResponseDto;
 import com.nhnacademy.frontend.common.service.BookService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,10 +12,14 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.FilterType;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.test.web.servlet.MockMvc;
 
 import java.time.LocalDate;
+import java.util.List;
 import java.util.Set;
 
 import static org.mockito.ArgumentMatchers.any;
@@ -48,24 +53,24 @@ public class AdminBookControllerTest {
                 .andExpect(view().name("admin/book/create-form"));
     }
 
-//    @Test
-//    void getBookList() throws Exception {
-//        SimpleBookResponseDto response1 = new SimpleBookResponseDto(1, "제목", "작가", 3000, 20, null, 1);
-//        SimpleBookResponseDto response2 = new SimpleBookResponseDto(2, "제목", "작가", 3000, 20, null, 1);
-//
-//        List<SimpleBookResponseDto> books = List.of(response1, response2);
-//        Page<SimpleBookResponseDto> page = new PageImpl<>(books);
-//
-//        when(bookService.getAllBooks(any(Pageable.class))).thenReturn(page);
-//
-//        mockMvc.perform(get("/admin/books"))
-//                .andExpect(status().isOk())
-//                .andExpect(view().name("admin/book/book-list"))
-//                .andExpect(model().attributeExists("books"))
-//                .andExpect(model().attributeExists("page"));
-//
-//        verify(bookService, times(1)).getAllBooks(any(Pageable.class));
-//    }
+    @Test
+    void getBookList() throws Exception {
+        SimpleBookResponseDto response1 = new SimpleBookResponseDto(1L, "제목", "작가", 3000, 20, null, 1L);
+        SimpleBookResponseDto response2 = new SimpleBookResponseDto(2L, "제목", "작가", 3000, 20, null, 1L);
+
+        List<SimpleBookResponseDto> books = List.of(response1, response2);
+        Page<SimpleBookResponseDto> page = new PageImpl<>(books);
+
+        when(bookService.getAllBooks(any(Pageable.class))).thenReturn(page);
+
+        mockMvc.perform(get("/admin/books"))
+                .andExpect(status().isOk())
+                .andExpect(view().name("admin/book/book-list"))
+                .andExpect(model().attributeExists("books"))
+                .andExpect(model().attributeExists("page"));
+
+        verify(bookService, times(1)).getAllBooks(any(Pageable.class));
+    }
 
     @Test
     void createBook_Success() throws Exception {
